requirements.txt
SQLAlchemy==1.3.19
python-dotenv==0.14.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
elasticsearch==6.3.1
redis==3.5.3
neo4j==4.0.2